requirements.txt
pysolr==3.9.0
pandas==2.1.4
pydantic==2.5.3
pyarrow==14.0.2
python-dotenv==1.0.0
richbench==1.0.3
